% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blip.R
\name{control_blip}
\alias{control_blip}
\title{Control arguments for doubly robust blip-learning}
\usage{
control_blip(blip_models = q_glm(~.), quantile_prob_threshold = NULL)
}
\arguments{
\item{blip_models}{Single element or list of V-restricted blip-models created
by \code{\link[=q_glm]{q_glm()}}, \code{\link[=q_rf]{q_rf()}} or similar functions.}

\item{quantile_prob_threshold}{Numeric vector. Quantile probabilities
for adaptively setting the threshold for the conditional average treatment effect.}
}
\value{
list of (default) control arguments.
}
\description{
\code{control_blip} sets the default control arguments
for doubly robust blip-learning, \code{type = "blip"}.
}
