% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drql.R
\name{control_drql}
\alias{control_drql}
\title{Control arguments for doubly robust Q-learning}
\usage{
control_drql(qv_models = q_glm(~.))
}
\arguments{
\item{qv_models}{Single element or list of V-restricted Q-models created
by \code{\link[=q_glm]{q_glm()}}, \code{\link[=q_rf]{q_rf()}} or similar functions.}
}
\value{
list of (default) control arguments.
}
\description{
\code{control_drql} sets the default control arguments
for doubly robust Q-learning, \code{type = "drql"}.
}
