% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqvl.R
\name{control_rqvl}
\alias{control_rqvl}
\title{Control arguments for QV-learning}
\usage{
control_rqvl(qv_models = q_glm(~.))
}
\arguments{
\item{qv_models}{Single element or list of V-restricted Q-models created
by \code{\link[=q_glm]{q_glm()}}, \code{\link[=q_rf]{q_rf()}}, \code{\link[=q_sl]{q_sl()}} or similar functions.}
}
\value{
list of (default) control arguments.
}
\description{
\code{control_rqvl} sets the default control arguments
for doubly robust V-restricted Q-learning, \code{type = "rqvl"}.
}
