% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_function.R
\name{fit_g_function}
\alias{fit_g_function}
\title{Fit single g-function}
\usage{
fit_g_function(history, g_model)
}
\arguments{
\item{history}{History object created by \code{\link[=get_history]{get_history()}}}

\item{g_model}{Propensity model/g-model created by \code{\link[=g_glm]{g_glm()}}, \code{\link[=g_rf]{g_rf()}}, \code{\link[=g_sl]{g_sl()}} or similar functions.}
}
\description{
\code{fit_g_function} is used to fit a single g-model
}
\examples{
library("polle")
### Simulating two-stage policy data
source(system.file("sim", "two_stage.R", package="polle"))
par0 <- c(gamma = 0.5, beta = 1)
d <- sim_two_stage(2e3, seed=1, par=par0)
pd <- policy_data(d,
                  action = c("A_1", "A_2"),
                  covariates = list(L = c("L_1", "L_2"),
                                    C = c("C_1", "C_2")),
                  utility = c("U_1", "U_2", "U_3"))
pd

history <- get_history(pd, stage = 1)
head(history$H)

g_function <- fit_g_function(
  history = get_history(pd, stage = 1),
  g_model = g_glm()
)
g_function
}
