% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{full_history}
\alias{full_history}
\title{Get the full history for a given stage}
\usage{
full_history(object, stage)
}
\arguments{
\item{stage}{stage number.}

\item{x}{Object of class \code{\link[=policy_data]{policy_data()}}.}
}
\value{
Object of class "history".
}
\description{
Get the full history for a given stage
}
