% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{full_stage_history}
\alias{full_stage_history}
\title{Get the full history for a given stage}
\usage{
full_stage_history(object, stage)
}
\arguments{
\item{object}{object of class \code{\link[=policy_data]{policy_data()}}.}

\item{stage}{stage number.}
}
\value{
Object of class "history".
}
\description{
Get the full history for a given stage
}
