% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_functions.R
\name{get_A}
\alias{get_A}
\title{Get the action vector A from a history object}
\usage{
get_A(history)
}
\arguments{
\item{history}{Object of class "history".}
}
\value{
Character vector.
}
\description{
Get the action vector A from a history object
}
