% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{get_event}
\alias{get_event}
\title{Get event indicators}
\usage{
get_event(object)
}
\arguments{
\item{object}{Object of class \link{policy_data} or \link{history}.}
}
\value{
\link[data.table:data.table]{data.table::data.table} with keys id and stage.
}
\description{
\code{get_event} returns the event indicators for given IDs
stages
}
