% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_functions.R
\name{get_H}
\alias{get_H}
\title{Get the history matrix H from a history object}
\usage{
get_H(history, vars = NULL)
}
\arguments{
\item{history}{Object of class "history".}

\item{vars}{Character vector. Subset of the history matrix.}
}
\value{
\link{data.table} H
}
\description{
Get the history matrix H from a history object
}
