% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{get_history_names.policy_data}
\alias{get_history_names.policy_data}
\title{Get history variable names}
\usage{
\method{get_history_names}{policy_data}(object, stage = NULL)
}
\arguments{
\item{stage}{Stage number. If NULL, the state/Markov-type history variable
names are returned.}

\item{policy_data}{Policy data object created by \code{\link[=policy_data]{policy_data()}}.}
}
\value{
Character vector.
}
\description{
\code{\link[=get_history_names]{get_history_names()}} returns the variable names of the history matrix
\eqn{H_k} for the given stage \eqn{k}. The function is useful when specifying
the design matrix for \code{\link[=g_model]{g_model()}} and \code{\link[=q_model]{q_model()}}.
}
\examples{
library("polle")
### Multiple stages:
source(system.file("sim", "multi_stage.R", package="polle"))
d3 <- sim_multi_stage(5e2, seed = 1)
# constructing policy_data object:
pd3 <- policy_data(data = d3$stage_data,
                   baseline_data = d3$baseline_data,
                   type = "long",
                   id = "id",
                   stage = "stage",
                   event = "event",
                   action = "A",
                   utility = "U")
pd3
# state/Markov type history variable names (H):
get_history_names(pd3)
# Full history variable names (H_k) at stage 2:
get_history_names(pd3, stage = 2)
}
