% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_functions.R
\name{get_id.history}
\alias{get_id.history}
\title{Get the ID vector from a history object}
\usage{
\method{get_id}{history}(object)
}
\arguments{
\item{history}{Object of class "history".}
}
\value{
Character vector.
}
\description{
Get the ID vector from a history object
}
