% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_functions.R
\name{get_id_stage.history}
\alias{get_id_stage.history}
\title{Get the IDs of the history object}
\usage{
\method{get_id_stage}{history}(object)
}
\arguments{
\item{object}{Object of class "history".}
}
\value{
\link{data.table} with variables id and stage.
}
\description{
Get the IDs of the history object
}
