% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{get_n.policy_data}
\alias{get_n.policy_data}
\title{Get Number of Observations}
\usage{
\method{get_n}{policy_data}(object)
}
\arguments{
\item{object}{Object of class \link{policy_data}.}
}
\value{
Integer.
}
\description{
\code{get_n} returns the number of observations in
the policy data object.
}
\examples{
### Two stages:
source(system.file("sim", "two_stage.R", package="polle"))
d2 <- sim_two_stage(5e2, seed=1)
# constructing policy_data object:
pd2 <- policy_data(d2,
                  action = c("A_1", "A_2"),
                  baseline = c("B"),
                  covariates = list(L = c("L_1", "L_2"),
                                    C = c("C_1", "C_2")),
                  utility = c("U_1", "U_2", "U_3"))
pd2

# getting the number of observations:
get_n(pd2)
}
