% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data.R
\name{new_policy_data}
\alias{new_policy_data}
\title{Create A Policy Data Object}
\usage{
new_policy_data(stage_data, baseline_data = NULL)
}
\arguments{
\item{stage_data}{A data.table on long format with required columns:
\itemize{
\item{id}
\item{stage: }{stage number of type integer.}
\item{event: }{0 indicating an action stage, 1 indicating a terminal stage and 2 indicating a censoring stage.}
\item{A: }{action variable of type character.}
\item{U: }{reward variable of type numeric.}
}
It is possible to add deterministic reward variables U_a for every action a in the action set, see details.
The remaining columns are considered state variables (X).}
}
\description{
\code{new_policy_data} creates an object of class policy_data.
}
\details{

}
