% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_eval_functions.R
\name{plot.policy_eval}
\alias{plot.policy_eval}
\title{Plot Diagnostics for a \code{policy_eval} Object}
\usage{
\method{plot}{policy_eval}(
  x,
  which = c(1),
  policy_data = NULL,
  stage = 1,
  history_variables = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class \link{policy_eval}}

\item{which}{A subset of the numbers 1:2
\itemize{
\item{1} Histogram of the influence curve terms
\item{2} Plot of the policy actions
}}

\item{policy_data}{Object of class \link{policy_data}}

\item{stage}{Stage number for plot 2}

\item{history_variables}{character vector of length 2 for plot 2}

\item{...}{Additional arguments passed to \code{\link[=hist]{hist()}} and \code{\link[=plot]{plot()}}}
}
\description{
Plot Diagnostics for a \code{policy_eval} Object
}
\examples{
d <- sim_two_stage(2e3, seed=1)
pd <- policy_data(d,
                  action = c("A_1", "A_2"),
                  baseline = "BB",
                  covariates = list(L = c("L_1", "L_2"),
                                    C = c("C_1", "C_2")),
                  utility = c("U_1", "U_2", "U_3"))

pe <- policy_eval(pd,
                  policy_learn = policy_learn())

plot(pe,
     which = c(1,2),
     policy_data = pd,
     stage = 1,
     history_variables = c("C", "L"))
}
