% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polle-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{IC}
\alias{estimate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lava}{\code{\link[lava:estimate]{estimate()}}, \code{\link[lava:IC]{IC()}}}
}}

