% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{state_history}
\alias{state_history}
\title{Get the state/Markov-type history for all stages}
\usage{
state_history(object)
}
\arguments{
\item{object}{object of class \code{\link[=policy_data]{policy_data()}}.}
}
\value{
Object of class "history".
}
\description{
Get the state/Markov-type history for all stages
}
