% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{state_stage_history}
\alias{state_stage_history}
\title{Get the state/Markov-type history for a given stage}
\usage{
state_stage_history(object, stage)
}
\arguments{
\item{object}{object of class \code{\link[=policy_data]{policy_data()}}.}

\item{stage}{stage number.}
}
\value{
Object of class "history".
}
\description{
Get the state/Markov-type history for a given stage
}
