% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data_functions.R
\name{utility.policy_data}
\alias{utility.policy_data}
\title{Get the Utility}
\usage{
\method{utility}{policy_data}(object)
}
\arguments{
\item{object}{Object of class \link{policy_data}.}
}
\value{
\link{data.table} with key id and numeric variable U.
}
\description{
\code{utility} returns the utility, i.e., the sum of the rewards,
for every observation in the policy data object.
}
\examples{
### Two stages:
source(system.file("sim", "two_stage.R", package="polle"))
d2 <- sim_two_stage(5e2, seed=1)
# constructing policy_data object:
pd2 <- policy_data(d2,
                  action = c("A_1", "A_2"),
                  baseline = c("B"),
                  covariates = list(L = c("L_1", "L_2"),
                                    C = c("C_1", "C_2")),
                  utility = c("U_1", "U_2", "U_3"))
pd2

# getting the utility:
head(utility(pd2))
}
