% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blip.R
\name{control_blip}
\alias{control_blip}
\title{Control arguments for doubly robust blip-learning}
\usage{
control_blip(blip_models = q_glm(~.))
}
\arguments{
\item{blip_models}{Single element or list of V-restricted blip-models created
by \code{\link[=q_glm]{q_glm()}}, \code{\link[=q_rf]{q_rf()}}, \code{\link[=q_sl]{q_sl()}} or similar functions.}
}
\value{
list of (default) control arguments.
}
\description{
\code{control_blip} sets the default control arguments
for doubly robust blip-learning, \code{type = "blip"}.
}
