% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_data.R
\name{policy_data}
\alias{policy_data}
\title{Create Policy Data Object}
\usage{
policy_data(
  data,
  baseline_data,
  type = "wide",
  action,
  covariates,
  utility,
  baseline = NULL,
  deterministic_rewards = NULL,
  id = NULL,
  stage = NULL,
  event = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{\link{data.frame} or \link{data.table}; see Examples.}

\item{baseline_data}{\link{data.frame} or \link{data.table}; see Examples.}

\item{type}{Character string. If "wide", \code{data} is considered to be on wide format.
Otherwise, \code{data} is considered to be on long format; see Examples.}

\item{action}{Action variable name(s). Character vector or character string.
\itemize{
\item{} A vector is valid for wide data. The length of the vector determines the number of stages (K).
\item{} A string is valid for single stage wide data or long data.
}}

\item{covariates}{Stage specific covariate name(s). Character vector or named list of character vectors.
\itemize{
\item{} A vector is valid for single stage wide data or long data.
\item{} A list is valid for multiple stage wide data. Each element of the list must be a character vector corresponding to the number of stages (K).
}}

\item{utility}{Utility/Reward variable name(s). Character string or vector.
\itemize{
\item{} A string is valid for long data and wide data with a single final utility.
\item{} A vector is valid for wide data with incremental rewards. Must be of length K+1; see Examples.
}}

\item{deterministic_rewards}{Deterministic reward variable name(s). Character vector or named list of character vectors.
\itemize{
\item{}
\item{}
}}
}
\description{
\code{policy_data} creates an object of class "policy_data".
}
\details{
Each observation has the sequential form
\deqn{O= {B, U_1, X_1, A_1, ..., U_K, X_K, A_K, U_{K+1}},}
for a possibly stochastic number of stages K.
\itemize{
\item{} B is a vector of baseline information.
\item{} U_k is the reward at stage k.
\item{} X_k is a vector summarizing the state at stage k.
\item{} A_k is the categorical action at stage k.
}
The utility is given by the sum of the rewards, i.e., \eqn{U = sum_{k = 1}^{K+1} U_k}.
}
\examples{

}
