% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_eval.R
\name{policy_eval}
\alias{policy_eval}
\title{Policy Evaluation}
\usage{
policy_eval(
  policy_data,
  policy = NULL,
  policy_learn = NULL,
  g_functions = NULL,
  g_models = g_glm(),
  g_full_history = FALSE,
  q_functions = NULL,
  q_models = q_glm(),
  q_full_history = FALSE,
  cross_fit = FALSE,
  M = 5,
  future_args = NULL,
  type = "dr"
)
}
\arguments{
\item{policy_data}{Policy data object}

\item{policy}{Policy object}

\item{policy_learn}{Policy learner object}

\item{g_functions}{Fitted g-model object}

\item{g_models}{Propensity model/g-model}

\item{g_full_history}{g-model: Full history or single stage/Markov history}

\item{q_functions}{Fitted Q-model object}

\item{q_models}{Outcome regression/Q-model}

\item{q_full_history}{Q-model: Full history or single stage/Markov history}

\item{cross_fit}{Should the evaluation be cross-fitted or not}

\item{M}{Number of folds}

\item{type}{Type of evaluation (dr/doubly robust, ipw/inverse propensity weighting, or/outcome regression)}
}
\description{
Policy evaluation function
}
