% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_eval.R
\name{policy_eval}
\alias{policy_eval}
\title{Policy Evaluation}
\usage{
policy_eval(
  policy_data,
  policy = NULL,
  policy_learner = NULL,
  g_functions = NULL,
  g_models = g_glm(),
  g_full_history = FALSE,
  q_functions = NULL,
  q_models = q_glm(),
  q_full_history = FALSE,
  M = 5,
  type = "dr",
  ...
)
}
\arguments{
\item{policy_data}{Policy data object}

\item{policy}{Policy object}

\item{g_models}{Propensity model object}

\item{g_full_history}{Full history or Markov}

\item{q_models}{Outcome regression/Q-model}

\item{q_full_history}{Full history or Markov}

\item{M}{Number of folds}

\item{type}{Type of model (dr, cv, ipw, or, ...)}

\item{...}{Additional arguments parsed to lower level functions}
}
\description{
Policy evaluation function
}
