% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_learn.R
\name{policy_learn}
\alias{policy_learn}
\title{Specify Policy Learner Object}
\usage{
policy_learn(
  type = "rql",
  alpha = 0,
  qv_models = NULL,
  qv_full_history = FALSE,
  policy_vars = NULL,
  policy_full_history = FALSE,
  L = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{type}{Character string. Type of policy learner method:
\itemize{
\item{} "rql": Realistic Quality/Q-learning
\item{} "rqvl": Realistic V-restricted Q-learning
\item{} "ptl": Policy Tree Learning
}}

\item{alpha}{probability limit for determining realistic actions.}
}
\description{
Specify Policy Learner Object
}
