% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polle-package.R
\docType{package}
\name{polle-package}
\alias{polle}
\alias{polle-package}
\title{polle: Policy Learning}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Package for learning and evaluating (subgroup) policies via doubly robust loss functions. Policy learning methods include doubly robust blip/conditional average treatment effect learning and sequential policy tree learning. Methods for (subgroup) policy evaluation include doubly robust cross-fitting and online estimation/sequential validation. See Nordland and Holst (2022) \doi{10.48550/arXiv.2212.02335} for documentation and references.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/AndreasNordland/polle/issues}
}

}
\author{
\strong{Maintainer}: Andreas Nordland \email{andreasnordland@gmail.com}

Authors:
\itemize{
  \item Klaus Holst \email{klaus@holst.it} (\href{https://orcid.org/0000-0002-1364-6789}{ORCID})
}

}
\keyword{internal}
