% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polle-package.R
\docType{package}
\name{polle-package}
\alias{polle-package}
\alias{polle}
\title{Policy Learning}
\description{
Framework evaluating and learning realistic policies based on
doubly robust loss functions.
}
\author{
Andreas Nordland (Maintainer) \href{mailto:andreas.nordland@gmail.com}{andreas.nordland@gmail.com}, Klaus Holst.
}
\keyword{package}
